requirements.txt
matplotlib==3.1.3
numpy==1.18.4
opencv-python==4.2.0.34
pandas==1.0.3
pep8==1.7.1
Pillow==7.1.2
scipy==1.4.1

